requirements.txt
colorama 
pycryptodomex 
pypsexec 
pywin32 
crypto 
pycryptodome